+++ src/router/index.js
@@ -83,6 +83,9 @@ const CustomersEdit = () => import('@/views/wny/customers/edit/CustomersEdit')
 const CustomersShow = () => import('@/views/wny/customers/show/CustomersShow')
 const AccountEdit = () => import('@/views/wny/customers/edit/AccountEdit')
 
+// SFC Customer Comments
+const CustomerComments = () => import('@/views/wny/customer-comments/Page');
+
 const Colors = () => import('@/views/theme/Colors')
 const Typography = () => import('@/views/theme/Typography')
 
@@ -505,6 +508,10 @@ export default new Router({
             {
               path: 'customers/:id/edit',
               component: CustomersEdit
+            },
+            {
+              path: 'customer-comments/:id',
+              component: CustomerComments
             }
           ]
         },