+++ src/views/wny/customer-comments/validation.js
@@ -1,0 +1,13 @@
+import {required} from 'vuelidate/lib/validators'
+
+export const validations = {
+  bodyComment: {
+    required
+  },
+  parentComment: {
+
+  },
+  replyParentId: {
+
+  }
+};