+++ src/views/wny/customers/edit/CustomersEdit.vue
@@ -61,8 +61,7 @@
               <b-form-input id="city"
                             v-model="$v.city.$model"
                             :class="status($v.city)"
-                            type="text"
-                            placeholder="City">
+                            type="text">
 
               </b-form-input>
             </b-form-group>
@@ -74,8 +73,7 @@
               <b-form-input id="line_1"
                             v-model="$v.line_1.$model"
                             :class="status($v.line_1)"
-                            type="text"
-                            placeholder="Mailing Address Line 1">
+                            type="text">
               </b-form-input>
             </b-form-group>
 
@@ -86,8 +84,7 @@
               <b-form-input id="line_2"
                             v-model="$v.line_2.$model"
                             :class="status($v.line_2)"
-                            type="text"
-                            placeholder="Mailing Address Line 2">
+                            type="text">
               </b-form-input>
             </b-form-group>