+++ src/views/wny/customers/index/Customers.vue
@@ -14,6 +14,7 @@
 
         <v-client-table :columns="columns" :data="data" :options="options" :theme="theme" id="dataTable">
           <p slot="actions" slot-scope="props">
+            <a :href="'#/admin/customer-comments/' + props.row.id" class="fa fa-comment-o action-icon"></a>
             <a :href="'#/admin/customers/' + props.row.id + '/edit'" class="icon-pencil action-icon"></a>
             <a class="icon-trash" v-on:click="deleteCustomer(props.row.id)" style="cursor: pointer"></a>
           </p>
@@ -28,102 +29,106 @@
 </template>
 
 <script>
-    import Vue from 'vue'
-    // import BCard from 'bootstrap-vue/es/components'
-    import {ClientTable, Event} from 'vue-tables-2'
-    import moment from "moment";
-    import axios from "../../../../backend/vue-axios/axios";
+  import Vue from 'vue'
+  import {ClientTable, Event} from 'vue-tables-2'
+  import moment from "moment";
+  import axios from "../../../../backend/vue-axios/axios";
+  import BCard from "bootstrap-vue/es/components/card/card";
+  import BCardHeader from "bootstrap-vue/es/components/card/card-header";
+  import BCardBody from "bootstrap-vue/es/components/card/card-body";
 
-    const API_URL = process.env.VUE_APP_API_URL;
-    Vue.use(ClientTable)
-    // Vue.use(BCard)
+  const API_URL = process.env.VUE_APP_API_URL;
+  Vue.use(ClientTable)
 
-    export default {
-        name: 'Customers',
-        components: {
-            ClientTable,
-            Event
+  export default {
+    name: 'Customers',
+    components: {
+      ClientTable,
+      Event,
+      BCard,
+      BCardHeader,
+      BCardBody
+    },
+    data: function () {
+      return {
+        columns: ['id', 'name', 'type', 'organization.name', 'created_at', 'updated_at', 'actions'],
+        data: [],
+        message: '',
+        success: false,
+        options: {
+          headings: {
+            id: 'ID',
+            name: 'Account Name',
+            type: 'Type',
+            'organization.name': 'Organization',
+            created_at: 'Created',
+            updated_at: 'Updated',
+            actions: 'Actions'
+          },
+          sortable: ['name', 'type', 'organization.name', 'created_date', 'updated_date'],
+          filterable: ['name', 'type', 'organization.name', 'created_date', 'updated_date'],
+          sortIcon: {base: 'fa', up: 'fa-sort-asc', down: 'fa-sort-desc', is: 'fa-sort'},
+          pagination: {
+            chunk: 5,
+            edge: false,
+            nav: 'scroll'
+          }
         },
-        data: function () {
-            return {
-                columns: ['id', 'name', 'type', 'organization.name', 'created_at', 'updated_at', 'actions'],
-                data: [],
-                message: '',
-                success: false,
-                options: {
-                    headings: {
-                        id: 'ID',
-                        name: 'Account Name',
-                        type: 'Type',
-                        'organization.name': 'Organization',
-                        created_at: 'Created',
-                        updated_at: 'Updated',
-                        actions: 'Actions'
-                    },
-                    sortable: ['name', 'type', 'organization.name', 'created_date', 'updated_date'],
-                    filterable: ['name', 'type', 'organization.name', 'created_date', 'updated_date'],
-                    sortIcon: {base: 'fa', up: 'fa-sort-asc', down: 'fa-sort-desc', is: 'fa-sort'},
-                    pagination: {
-                        chunk: 5,
-                        edge: false,
-                        nav: 'scroll'
-                    }
-                },
-                useVuex: false,
-                theme: 'bootstrap4',
-                template: 'default'
-            }
-        },
-        methods: {
-            deleteCustomer: function (customerId) {
-                let headers = {
-                    headers: {
-                        'Accept': 'application/json',
-                        'Authorization': 'Bearer ' + localStorage.token
-                    }
-                };
+        useVuex: false,
+        theme: 'bootstrap4',
+        template: 'default'
+      }
+    },
+    methods: {
+      deleteCustomer: function (customerId) {
+        let headers = {
+          headers: {
+            'Accept': 'application/json',
+            'Authorization': 'Bearer ' + localStorage.token
+          }
+        };
 
-                axios.delete(API_URL + '/customers/' + customerId, headers)
-                    .then(request => this.customerDeletingSuccessful(request))
-                    .catch((request) => this.customerDeletingFailed(request));
-            },
-            customerDeletingSuccessful(req) {
-                this.errors = false;
-                this.error = false;
-                this.flash('The Customer is deleted.', 'success');
+        axios.delete(API_URL + '/customers/' + customerId, headers)
+             .then(request => this.customerDeletingSuccessful(request))
+             .catch((request) => this.customerDeletingFailed(request));
+      },
+      customerDeletingSuccessful(req) {
+        this.errors = false;
+        this.error = false;
+        this.flash('The Customer is deleted.', 'success');
 
-                this.downloadData();
-            },
-            customerDeletingFailed(req) {
-                this.errors = false;
-                this.error = 'Customer Deleting failed! ' + req;
-            },
+        this.downloadData();
+      },
+      customerDeletingFailed(req) {
+        this.errors = false;
+        this.error = 'Customer Deleting failed! ' + req;
+      },
 
-            downloadData() {
-                let headers = {
-                    headers: {
-                        'Accept': 'application/json',
-                        'Authorization': 'Bearer ' + localStorage.token
-                    }
-                };
+      downloadData() {
+        let headers = {
+          headers: {
+            'Accept': 'application/json',
+            'Authorization': 'Bearer ' + localStorage.token
+          }
+        };
 
-                axios.get(API_URL + '/customers', headers)
-                    .then(response => {
-                        this.data = response.data.data;
-                        this.message = response.data.message;
-                        this.success = response.data.success;
-                        console.log(this.message);
-                        console.log(this.success);
-                        console.log("data: " + this.data);
-                        console.log("start_date: " + this.data[1].start_date);
-                    })
-                    .catch(error => console.log(error));
-            }
-        },
-        mounted() {
-            this.downloadData();
-        }
-    };
+        axios.get(API_URL + '/customers', headers)
+             .then(response => {
+               this.data = response.data.data;
+               this.message = response.data.message;
+               this.success = response.data.success;
+               console.log(this.message);
+               console.log(this.success);
+               console.log("data: " + this.data);
+               console.log("start_date: " + this.data[1].start_date);
+             })
+             .catch(error => console.log(error));
+      }
+    },
+    mounted() {
+      this.downloadData();
+    }
+  };
 </script>
 
 <style lang="scss">