+++ CHANGELOG.md
@@ -3,8 +3,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Changed
-- Customer. Create, Edit. Add "&" in the available signs in the field Customer Account Name.
-- Customer. Create, Edit. Field "City" is implemented. Fields "City", "Line 1" are required.
+- Customers. Create, Edit. Add "&" in the available signs in the field Customer Account Name.
+- Customers. Create, Edit. Field "City" is implemented. Fields "City", "Line 1" are required.
+- Customers. Create, Edit. Placeholders are deleted.
 
 ## [0.17.0] - 2019-11-21
 ### Changed